requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
apache-airflow-providers-postgres

